requirements.txt
google-genai==1.7.0
python-dotenv==1.1.0
langgraph==0.3.21
langchain-google-genai==2.1.2
langgraph-prebuilt==0.1.7